+++ elmyra.ip.access.epo/CHANGES.rst
@@ -9,6 +9,7 @@ development
 - ui, storage: introduce multi-tenancy by "context" query parameter
 - ui, storage: fix localForage keys() bug
 - ui, storage: add jquery.base64 library
+- ui, storage: add stripped-down dataurl library
 
 0.19.0
 ======